moved audio analysis into its own module
